requirements.txt
Flask==3.0.0
flask-login==0.6.3
flask-wtf==1.2.1
WTForms==3.1.1
Flask-SQLAlchemy==3.1.1
passlib==1.7.4
python-dotenv==1.0.0

